Add marshaling and unmarshaling support
This commit adds support for marshaling and unmarshaling Data to either
JSON or YAML.  The calculated fields--variance, sample_variance, std_dev,
and sample_std_dev--are used to calculate the internal "m2" value on
deserialization, but note that this calculation will result in the
introduction of slight errors.  The exact calculated fields to include may
be specified by a bit mask in the Data.Flags field; if it is 0, all
calculated fields are included.